#970 move general fixtures to higher conftest
