+++ tests/system_tests/experiment_plans/test_fgs_plan.py
@@ -43,11 +43,6 @@ def params():
 
 
 @pytest.fixture
-def RE():
-    return RunEngine({})
-
-
-@pytest.fixture
 def fgs_composite():
     composite = FlyScanXRayCentreComposite(
         attenuator=i03.attenuator(),
